Refactor all route functions into application.py
